fix benchmark

+++ frame/nfts/src/lib.rs
@@ -77,12 +77,12 @@ pub mod pallet {
 
 	#[cfg(feature = "runtime-benchmarks")]
 	pub trait BenchmarkHelper<CollectionId, ItemId> {
-		fn collection(i: u16) -> CollectionId;
+		fn collection(i: u32) -> CollectionId;
 		fn item(i: u16) -> ItemId;
 	}
 	#[cfg(feature = "runtime-benchmarks")]
-	impl<CollectionId: From<u16>, ItemId: From<u16>> BenchmarkHelper<CollectionId, ItemId> for () {
-		fn collection(i: u16) -> CollectionId {
+	impl<CollectionId: From<u32>, ItemId: From<u16>> BenchmarkHelper<CollectionId, ItemId> for () {
+		fn collection(i: u32) -> CollectionId {
 			i.into()
 		}
 		fn item(i: u16) -> ItemId {
